\name{evstat-package}
\alias{evstat-package}
\docType{package}
\title{
Extreme Value Statistics
}
\description{
The package evstat has a set of functions  
}
\details{
An overview of how to use the package, including the most important functions
}
\author{
Frederico Caeiro
}
\references{
%Optionally other standard keywords, one per line, from file KEYWORDS  in the R documentation directory ~~
Beirlant, J., Goegebeur, Y.,  Segers, J.  and  Teugels, J. (2004). \emph{Statistics of Extremes: Theory and Applications}, Wiley, England.

Beirlant, J., Caeiro, F.  and Gomes, M.I.  (2012). An overview and open researh topics in statistics of univariate extremes. \emph{Revstat} \bold{10}:1, 1--31. 

Gomes, M.I. and Guillou, A. (2014). Extreme Value Theory and Statistics of Univariate Extremes: A Review. \emph{International Statistical Review}, accepted.

Reiss, R.-D. and Thomas, M. (2007). \emph{Statistical Analysis of Extreme Values, with Application to Insurance, Finance, Hydrology and Other Fields},  3rd edition, Birkhauser Verlag, Basel.
}
\keyword{ package }
%\seealso{
% Optional links to other man pages, e.g. 
% \code{\link[<pkg>:<pkg>-package]{<pkg>}} 
%}
%\examples{
%}
